+++ grafanimate/scenarios.py
@@ -1,17 +1,15 @@
 #  -*- coding: utf-8 -*-
-# (c) 2018 Andreas Motl <andreas@hiveeyes.org>
+# (c) 2018-2021 Andreas Motl <andreas@hiveeyes.org>
 # License: GNU Affero General Public License, Version 3
 import logging
 from datetime import datetime
 
-from grafanimate.animations import SequentialAnimation
-from grafanimate.grafana import GrafanaWrapper
-from grafanimate.model import NavigationFlavor
+from grafanimate.model import NavigationFlavor, AnimationStep
 
 logger = logging.getLogger(__name__)
 
 
-class AnimationScenario:
+class BuiltinScenarios:
     """
     Run different animation scenarios/sequences.
 
@@ -25,20 +23,6 @@ class AnimationScenario:
     - "expand" will use a fixed start time and stepwise expand the end time
     """
 
-    def __init__(self, grafana: GrafanaWrapper, storage, dashboard_uid=None, target=None, options=None):
-
-        self.grafana = grafana
-        self.storage = storage
-        self.dashboard_uid = dashboard_uid
-
-        # Dispatch output to different target formats.
-        # FIXME: Not implemented yet.
-        self.target = target
-
-        # Start the engines.
-        self.engine = SequentialAnimation(grafana=grafana, dashboard_uid=dashboard_uid, options=options)
-        self.engine.start()
-
     def playdemo(self):
         """
         Run demo on `play.grafana.org`.
@@ -48,80 +32,79 @@ class AnimationScenario:
             grafanimate --grafana-url=https://play.grafana.org/ --dashboard-uid=000000012 --scenario=playdemo
         """
         logger.info('Running scenario playdemo')
-
-        results = self.engine.run(dtstart=datetime(2021, 11, 14, 2, 0, 0), dtuntil=datetime(2021, 11, 14, 4, 16, 36), interval='5min', flavor=NavigationFlavor.EXPAND)
-        self.storage.save_items(results)
+        return AnimationStep(dtstart=datetime(2021, 11, 14, 2, 0, 0), dtuntil=datetime(2021, 11, 14, 4, 16, 36), interval='5min', flavor=NavigationFlavor.EXPAND)
 
     def ldi_all(self):
+        """
+        Luftdaten.info, all
+        """
         logger.info('Running scenario ldi_all')
 
-        # LDI, ramp-up
-        results = self.engine.run(dtstart=datetime(2015, 10, 1), dtuntil=datetime(2017, 1, 1), interval='monthly')
-        self.storage.save_items(results)
+        return [
 
-        # LDI, growth
-        results = self.engine.run(dtstart=datetime(2017, 1, 1), dtuntil=datetime.now(), interval='weekly')
-        self.storage.save_items(results)
+            # LDI, ramp-up
+            AnimationStep(dtstart=datetime(2015, 10, 1), dtuntil=datetime(2017, 1, 1), interval='monthly'),
+
+            # LDI, growth
+            AnimationStep(dtstart=datetime(2017, 1, 1), dtuntil=datetime.now(), interval='weekly'),
+
+        ]
 
     def ldi_with_gaps(self):
+        """
+        Luftdaten.info, growth
+        """
         logger.info('Running scenario ldi_with_gaps')
 
-        # LDI, ramp-up
-        results = self.engine.run(dtstart=datetime(2015, 10, 1), dtuntil=datetime(2017, 1, 1), interval='monthly')
-        self.storage.save_items(results)
+        return [
 
-        # LDI, growth, with gap at 2018-04-29 - 2018-12-20
-        # TODO: Detect empty data from datasource through Grafana Sidecar and skip respective images.
-        results = self.engine.run(dtstart=datetime(2017, 1, 1), dtuntil=datetime(2018, 6, 5), interval='weekly')
-        self.storage.save_items(results)
+            # LDI, ramp-up
+            AnimationStep(dtstart=datetime(2015, 10, 1), dtuntil=datetime(2017, 1, 1), interval='monthly'),
 
-        results = self.engine.run(dtstart=datetime(2018, 12, 20), dtuntil=datetime.now(), interval='weekly')
-        self.storage.save_items(results)
+            # LDI, growth, with gap at 2018-04-29 - 2018-12-20
+            # TODO: Detect empty data from datasource through Grafana Sidecar and skip respective images.
+            AnimationStep(dtstart=datetime(2017, 1, 1), dtuntil=datetime(2018, 6, 5), interval='weekly'),
+
+            # LDI, until now
+            AnimationStep(dtstart=datetime(2018, 12, 20), dtuntil=datetime.now(), interval='weekly'),
+        ]
 
     def ldi_nye_2017_2018(self):
+        """
+        LDI, New Year's Eve 2018
+        """
         logger.info('Running scenario ldi_nye_2017_2018')
-
-        # LDI, New Year's Eve 2018
-        results = self.engine.run(dtstart=datetime(2017, 12, 31, 21, 0, 0), dtuntil=datetime(2018, 1, 1, 4, 0, 0), interval='10min')
-        self.storage.save_items(results)
+        return AnimationStep(dtstart=datetime(2017, 12, 31, 21, 0, 0), dtuntil=datetime(2018, 1, 1, 4, 0, 0), interval='10min')
 
     def ldi_nye_2018_2019(self):
+        """
+        LDI, New Year's Eve 2018/2019
+        """
         logger.info('Running scenario ldi_nye_2018_2019')
-
-        # LDI, New Year's Eve 2018/2019
-        results = self.engine.run(dtstart=datetime(2018, 12, 31, 15, 0, 0), dtuntil=datetime(2018, 12, 31, 20, 0, 0), interval='30min')
-        self.storage.save_items(results)
-
-        results = self.engine.run(dtstart=datetime(2018, 12, 31, 20, 0, 0), dtuntil=datetime(2018, 12, 31, 23, 0, 0), interval='10min')
-        self.storage.save_items(results)
-
-        results = self.engine.run(dtstart=datetime(2018, 12, 31, 23, 0, 0), dtuntil=datetime(2019, 1, 1, 1, 0, 0), interval='5min')
-        self.storage.save_items(results)
-
-        results = self.engine.run(dtstart=datetime(2019, 1, 1, 1, 0, 0), dtuntil=datetime(2019, 1, 1, 4, 0, 0), interval='10min')
-        self.storage.save_items(results)
-
-        results = self.engine.run(dtstart=datetime(2019, 1, 1, 4, 0, 0), dtuntil=datetime(2019, 1, 1, 9, 0, 0), interval='30min')
-        self.storage.save_items(results)
+        return [
+            AnimationStep(dtstart=datetime(2018, 12, 31, 15, 0, 0), dtuntil=datetime(2018, 12, 31, 20, 0, 0), interval='30min'),
+            AnimationStep(dtstart=datetime(2018, 12, 31, 20, 0, 0), dtuntil=datetime(2018, 12, 31, 23, 0, 0), interval='10min'),
+            AnimationStep(dtstart=datetime(2018, 12, 31, 23, 0, 0), dtuntil=datetime(2019, 1, 1, 1, 0, 0), interval='5min'),
+            AnimationStep(dtstart=datetime(2019, 1, 1, 1, 0, 0), dtuntil=datetime(2019, 1, 1, 4, 0, 0), interval='10min'),
+            AnimationStep(dtstart=datetime(2019, 1, 1, 4, 0, 0), dtuntil=datetime(2019, 1, 1, 9, 0, 0), interval='30min'),
+        ]
 
     def cdc_maps(self):
+        """
+        DWD CDC, temperatur-sonne-and-niederschlag-karten
+        """
         logger.info('Running scenario cdc_maps')
-
-        # CDC, temperatur-sonne-and-niederschlag-karten
-        results = self.engine.run(dtstart=datetime(2018, 3, 6, 5, 0, 0), dtuntil=datetime(2018, 3, 10, 23, 59, 59), interval='hourly')
+        return AnimationStep(dtstart=datetime(2018, 3, 6, 5, 0, 0), dtuntil=datetime(2018, 3, 10, 23, 59, 59), interval='hourly')
 
         # Short sequence, for debugging processes.
-        #results = self.engine.run(dtstart=datetime(2018, 3, 6, 5, 0, 0), dtuntil=datetime(2018, 3, 6, 6, 59, 59), interval='hourly')
-
-        self.storage.save_items(results)
+        return AnimationStep(dtstart=datetime(2018, 3, 6, 5, 0, 0), dtuntil=datetime(2018, 3, 6, 6, 59, 59), interval='hourly')
 
     def uba_ldi_dwd_maps(self):
+        """
+        Labor: Studio / UBA/LDI/DWD-Studio [dev!]
+        """
         logger.info('Running scenario uba_ldi_dwd_maps')
-
-        # Labor: Studio / UBA/LDI/DWD-Studio [dev!]
-        results = self.engine.run(dtstart=datetime(2018, 10, 6, 5, 0, 0), dtuntil=datetime(2018, 10, 10, 23, 59, 59), interval='hourly')
-
-        self.storage.save_items(results)
+        return AnimationStep(dtstart=datetime(2018, 10, 6, 5, 0, 0), dtuntil=datetime(2018, 10, 10, 23, 59, 59), interval='hourly')
 
     def ir_sensor_svg_pixmap(self):
         """
@@ -129,6 +112,4 @@ class AnimationScenario:
         dtuntil: 2018-08-14 03:16:36
         """
         logger.info('Running scenario ir_sensor_svg_pixmap')
-
-        results = self.engine.run(dtstart=datetime(2018, 8, 14, 3, 16, 0), dtuntil=datetime(2018, 8, 14, 3, 16, 36), interval='secondly')
-        self.storage.save_items(results)
+        return AnimationStep(dtstart=datetime(2018, 8, 14, 3, 16, 0), dtuntil=datetime(2018, 8, 14, 3, 16, 36), interval='secondly')