+++ CHANGES.rst
@@ -12,6 +12,7 @@ in progress
 - Add compatibility with Grafana 6, 7 and 8
 - Add possibility to login to protected Grafana instances
 - Add possibility to control the navigation flavor (window, expand)
+- Introduce data model for animation scenarios
 
 
 2019-02-04 0.5.5